src/agents/document_processor.py
# Conceptual flow for the Document Processor node in LangGraph
import pytesseract
from PIL import Image
import fitz  # PyMuPDF

def document_processor_node(state):
    """
    A LangGraph node that takes a file path from the state,
    performs OCR, and updates the state with extracted info.
    """
    file_path = state["input_document"]

    # 1. If it's a PDF, convert first page to image using PyMuPDF
    if file_path.endswith(".pdf"):
        doc = fitz.open(file_path)
        pix = doc[0].get_pixmap(dpi=300)  # Render first page
        img = Image.frombytes("RGB", [pix.width, pix.height], pix.samples)
    else:
        img = Image.open(file_path)

    # 2. Perform OCR with Tesseract
    #    Use config for better accuracy on structured docs
    custom_config = r'--oem 3 --psm 6'  # Assume uniform text block
    extracted_text = pytesseract.image_to_string(img, config=custom_config)

    # 3. In a real agent, you'd use an LLM to parse the text
    #    e.g., model.invoke(f"Extract weight, dimensions, and special_handling from:\n{extracted_text}")
    
    # 4. Update the state with the new info
    return {
        "document_text": extracted_text,
        "package_weight": 15.5,  # Placeholder from LLM parsing
        "special_handling": ["fragile", "refrigerated"] # Placeholder
    }